Fixed several bugs that showed up in Leopard but were present still in Tiger. Mostly just bugs dealing with calling super release too early.
Changed callback method for Trivia Board transitions to use a normal delegate.

Fixed crazy bug caused by TriviaTimers not getting stopped and so they called after we thought they were done (the extra retain from NSTimer was killer). This bug was present in Tiger but happened too infrequently to debug. It happened so often in Leopard that debugging was easy and necessary.

- updated version to 1.0.2